model modified, create meeting implemented

+++ src/main/java/mag/joinus/model/LatLng.java
@@ -41,7 +41,7 @@ public class LatLng {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
